scripts/requirements.txt
openpyxl
requests
pyyaml